requirements.txt
PyYAML~=5.4.1
PySimpleGUI~=4.39.1